added user table relationship

+++ controllers/lists/lists.js
@@ -1,36 +1,38 @@
 const listModel = require('../../models/lists');
 
 exports.getAllLists = (req, res) => {
-    listModel.find({}).then((lists) => {
+    const id = req.params.userId;
+    listModel.find({ author: id }).then((list) => {
         res.json({
             msg: 'Successfully gotten all lists',
-            data: lists
+            data: list
         });
     }).catch(error => {
         res.status(400).json({ err: error });
     })
 }
 
-exports.createList = (req, res) => {
+exports.createList = async(req, res) => {
     const title = req.body.title;
+    const author = req.body.author;
     const newList = new listModel({
-        title
+        title,
+        author
     });
-    try {
-        newList.save().then((list) => {
-            res.json({
-                msg: 'List added successfully',
-                data: list
-            })
+    newList.save().then((list) => {
+        res.json({
+            msg: 'List added successfully',
+            data: list
         })
-    } catch (error) {
+    }).catch(error => {
         res.status(400).json({ err: error });
-    }
+    });
 }
 
 exports.updateList = (req, res) => {
     const listId = req.params.id;
-    listModel.findByIdAndUpdate({ _id: listId }, {
+    const authorId = req.params.authorId;
+    listModel.findByIdAndUpdate({ author: authorId, _id: listId }, {
         $set: req.body
     }).then(() => {
         res.json({
@@ -43,12 +45,13 @@ exports.updateList = (req, res) => {
 
 exports.deleteList = (req, res) => {
     const listId = req.params.id;
-    if (!listId) {
+    const authorId = req.params.authorId;
+    if (!listId || !authorId) {
         res.json({
             msg: "list id is required"
         })
     }
-    listModel.findByIdAndRemove({ _id: listId }).then((removedList) => {
+    listModel.findByIdAndRemove({ author: authorId, _id: listId }).then((removedList) => {
         res.json({
             msg: "List removed successfully",
             data: removedList
